update method %n

+++ roll4init.rb
@@ -445,15 +445,20 @@ bot.message(start_with: "%n") do |event|; #next initiative revealed
         say = say + sqlCode;
       # Execute SQL CODE 
         result = conn.exec(sqlCode);
-        say = say + "\n" + result.ntuples.to_s;
-        say = say + "\n" + result.inspect;
-        result.each do |item|
-            say = say + "\n" + item.inspect;
-            theID = item.fetch("id").to_s;
-            say = say + "\n> the ID " + theID;
-            sqlCode = "DELETE FROM activeInit WHERE id = " + theID + ";";
-            conn.exec(sqlCode);
-        end;
+        goNogo = result.ntuples.to_i;
+        if goNogo != 0 then;
+           say = say + "\n" + result.inspect;
+           result.each do |item|
+                 say = say + "\n" + item.inspect;
+                 theID = item.fetch("id").to_s;
+                 say = say + "\n> the ID " + theID;
+                 sqlCode = "DELETE FROM activeInit WHERE id = " + theID + ";";
+                 conn.exec(sqlCode);
+           end;
+         else;
+           say = say + "\n ** END OF THE COMBAT ROUND**";
+         end;
+        
   conn.close;
   event.respond say;
 end;
